Add check for refunds array if it is not undefined
In a shop for a client, I was having an issue when using mollie credit card payment that the order items in the order details was not showing up in the admin section. The error was giving an "undefined" error at this point when reading the length property. Adding an additional check here for refunds array if it is available before checking for length solves the problem.