Update price check to include zero value (#1161)
* Update price check to include zero value

When the cart is empty the price is zero and because of that the $availableMolliePaymetnts returns empty. This cause issue where you won't be able load the payments using the SW PaymentMethodRoute. By included the zero condition you will get the original data.

* Update CHANGELOG with recent fixes and updates

Updated changelog with fixes and updates for versions 4.20.0 and 4.20.1.

* Update CHANGELOG for version 4.20.1

Updated changelog for version 4.20.1 with bug fixes and documentation corrections.