Update the ApplePayDomainVerificationService for compatibility reasons. (#1143)
* update ApplePayDomainVerificationService to be compatible with more flysystem adapters

using the older fileExists function is better compatible with the flysystem-async-aws-s3

also added a config array to the write function as autodetermination of the mimtype for the apple-developer-merchantid-domain-association file is not working well.

* Add error handling to ApplePay domain verification

Wrapped filesystem operations in a try-catch block to handle potential exceptions during file deletion and writing in ApplePayDomainVerificationService. Also updated method to use has() instead of fileExists().